requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
cycler==0.10.0
idna==2.5
matplotlib==2.0.2
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyasn1==0.2.3
pycparser==2.17
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.2
six==1.10.0
urllib3==1.21.1
